GUA:470: Rename getLocalStackConfig, getLocalStackKmsConfig & add AwsConfig
We'll need to make something very similar to this for SNS in the next commit.

Let's avoid any potential confusion or overlap by giving this a more specific
name.

Also the SNS config is likely to share much of the AwsConfig payload, rather
than creating a nearly identical setup let's extract what we'll need from here
and allow it to be shared.