GUA-462: Query SerivceList Store, filter and present services used
This commit breaks the pattern of this repo in focusing on functions, instead
it starts to use constants with arrow syntax, mostly as it's a pattern that
we've been using to good effect in our backend lambdas, and fits with the
GOV.UK Account Team's preference to the AirBnB Style guide and functional
Typescript. This is a break and introduces an inconsistency between existing
code and our new code here. That should be given due consideration during
review.

In this case I thought it was appropraite as the functionality is entirely new,
and we're not looking to mirror anything existing, querying the service store
is a brand new piece of functionality.

We should decide in this review if this is a direction we should continue in,
or if I should refactor this.

Additionally we will need to maintain an allowed list of both services that we
wish to show, but which of the two lists we'd want them to appear in. For that
I will need actual ClientIDs.

We may need to come up with something my dynamic than this so we can represent
ClientIDs from the various Environments, for now, I'll just hard code strings
for testing and we can return to this in a later commit.